Move photo cropper into its own file
