+++ AgriHealth-Alert-main/app/src/main/java/com/android/agrihealth/ui/office/ManageOfficeScreen.kt
@@ -301,6 +301,7 @@ fun EditableProfilePicture(
   }
 }
 
+// TODO Make the profile picture square for offices
 @Composable
 fun UploadRemoveOfficePhotoSection(
     isOwner: Boolean = false,
